+++ CHANGELOG.md
@@ -1,1 +1,9 @@
-Initial Release v0.1.0
+# Version 0.1.1
+- Added option to fully hide crosshair (except for menus).
+- Added option to use the Windows Cursor as the crosshair.
+- Added new hot spot selection options for better compatibility with certain crosshair types:<br>
+`Center Left, Center Right, Top Center, Bottom Center`
+
+- Minor code cleanup.
+- Adjusted README.md
+- Added AGPL 3.0 License to project files.